Protection Routes

+++ controllers/videoController.js
@@ -59,7 +59,7 @@ export const videoDetail = async (req, res) => {
 
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
+
     res.render("VideoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -72,7 +72,13 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator) {
+      if (video.creator !== req.user.id) {
+        throw Error();
+      } else {
+        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+      }
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -95,7 +101,15 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator) {
+      if (video.creator !== req.user.id) {
+        throw Error();
+      } else {
+        await Video.findOneAndRemove({ _id: id });
+      }
+    }
+
     // eslint-disable-next-line no-empty
   } catch (error) {}
   res.redirect(routes.home);
